requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
Django==4.2.29
django-environ==0.13.0
django-inertia==1.3.0
django-vite==3.1.0
dotty_dict==1.3.0
fonttools==4.62.1
idna==3.11
inertia-django==1.2.0
kiwisolver==1.5.0
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
setuptools==82.0.1
setuptools-scm==10.0.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
vcs-versioning==1.1.1
whitenoise==6.12.0
